chore: move web3-kit to web3-shared-base
