--- packages/web3-kit/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@masknet/web3-kit",
-  "private": true,
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts"
-}
