--- packages/web3-kit/src/index.ts
@@ -1,2 +1,0 @@
-export * from './helpers'
-export * from './types'
