+++ packages/web3-shared/base/src/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './constants'
 export * from './lookupTable'
+export * from './types'
