+++ packages/web3-shared/base/src/helpers/lookupTable.ts
