+++ packages/web3-shared/base/src/index.ts
@@ -1,3 +1,4 @@
 export * from './hooks'
 export * from './utils'
 export * from './proxy'
+export * from './helpers'
