+++ packages/web3-shared/flow/constants/index.ts
@@ -1,6 +1,6 @@
 import Auth from '@masknet/web3-constants/flow/auth.json'
 import Token from '@masknet/web3-constants/flow/token.json'
-import { hookTransform, transform } from '@masknet/web3-kit'
+import { hookTransform, transform } from '@masknet/web3-shared-base'
 import { ChainId } from '../types'
 
 export const getAuthConstants = transform(ChainId, Auth)
