+++ packages/web3-shared/flow/tsconfig.json
@@ -6,5 +6,5 @@
     "tsBuildInfoFile": "../dist/flow.tsbuildinfo"
   },
   "include": ["./", "./**/*.json"],
-  "references": [{ "path": "../../web3-kit" }]
+  "references": [{ "path": "../../web3-shared/base" }]
 }
