+++ packages/web3-shared/solana/constants/index.ts
@@ -1,5 +1,5 @@
 import Token from '@masknet/web3-constants/solana/token.json'
-import { hookTransform, transform } from '@masknet/web3-kit'
+import { hookTransform, transform } from '@masknet/web3-shared-base'
 import { ChainId } from '../types'
 
 export const getTokenConstants = transform(ChainId, Token)
