+++ packages/web3-shared/solana/tsconfig.json
@@ -6,5 +6,5 @@
     "tsBuildInfoFile": "../dist/solana.tsbuildinfo"
   },
   "include": ["./", "./**/*.json"],
-  "references": [{ "path": "../../web3-kit" }]
+  "references": [{ "path": "../../web3-shared/base" }]
 }
