+++ tsconfig.json
@@ -79,7 +79,6 @@
       "@masknet/theme": ["./packages/theme/src"],
       // @masknet/web3-constants is not listed. It is not a TS project (only contains JSON files).
       // @masknet/web3-contracts is not listed. It is not a TS project (only contains generated JSON and .d.ts files).
-      "@masknet/web3-kit": ["./packages/web3-kit/src"],
       "@masknet/web3-providers": ["./packages/web3-providers/src"],
       "@masknet/web3-shared-base": ["./packages/web3-shared/base/src"],
       "@masknet/web3-shared-evm": ["./packages/web3-shared/evm/src"],
